requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3        # newer version with better compatibility
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7           # pin to a version that works with Flask-Login 0.6.3